crypto-key-api/requirements.txt
base58==2.1.1
blinker==1.9.0
bloom-filter2==2.0.0
click==8.2.1
ecdsa==0.19.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycryptodome==3.23.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
